Split analyze_issues.py (3546 LOC) into 4 cohesive modules
The god-module is decomposed by responsibility into sibling modules
under python/larch/issue/:
- _util.py: constants, parse_iso, load_issues, issue_number
- _report.py: coverage_stats, categorize, growth_chart, reviewer_effectiveness
- _oos.py: OOS parsing, fate scoring, iter_filed_oos_records
- _ground_truth.py: GroundTruth* types and voter calibration

analyze_issues.py becomes a thin wrapper that re-exports all public
symbols for backward compatibility.

Also updates complexity-baseline.json (regen via make regen-complexity-baseline)
and ruff.toml (new per-file C901/PLR suppression entries for the new modules).
