+++ python/ruff.toml
@@ -149,9 +149,28 @@ ignore = [
     "PLR0913",
     "PLR0915",
 ]
-"analyze_issues.py" = [
+"_ground_truth.py" = [
     "C901",
     "PLR0911",
+    "PLR0912",
+    "PLR0913",
+    "PLR0915",
+]
+"_oos.py" = [
+    "C901",
+    "PLR0911",
+    "PLR0912",
+    "PLR0913",
+    "PLR0915",
+]
+"_report.py" = [
+    "C901",
+    "PLR0912",
+    "PLR0913",
+    "PLR0915",
+]
+"analyze_issues.py" = [
+    "C901",
     "PLR0912",
     "PLR0913",
     "PLR0915",
